Remove redundancy of javaapi pkg
